+++ src/components/common/Icons/QuestionIcon.vue
@@ -1,0 +1,29 @@
+<template>
+  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 46 46">
+    <path
+      d="M23 45c12.15 0 22-9.85 22-22S35.15 1 23 1 1 10.85 1 23s9.85 22 22 22Z"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-miterlimit="10"
+      stroke-linecap="square"
+    />
+    <path
+      d="M4.204 15.774h2.45v-.126c.04-2.608.693-3.744 2.489-4.9C10.939 9.622 12 8.003 12 5.678 12 2.399 9.673 0 6.082 0 2.776 0 .153 2.103 0 5.679h2.571c.153-2.482 1.837-3.492 3.51-3.492 1.919 0 3.47 1.304 3.47 3.366 0 1.672-.929 2.87-2.122 3.617-2 1.252-3.194 2.471-3.225 6.478zM5.51 22c1.01 0 1.837-.852 1.837-1.893s-.827-1.893-1.837-1.893-1.837.852-1.837 1.893S4.5 22 5.51 22Z"
+      fill="currentColor"
+      transform="translate(17 12)"
+    />
+  </svg>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({});
+</script>
+
+<style lang="scss" scoped>
+svg {
+  width: 1em;
+  height: 1em;
+}
+</style>