+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixes
 
+- Improved partial swap result
+
 ### Perfomance
 
 ## 0.1.2 - 2021-09-29